Start of conversion to entity-component-system architecture
Example here is turning the rendering of everything to be done by a
rendering system that uses the RenderComponent and PositionComponent.
These components are stored in a dictionary identified by a key the
identifies the associated entity.

Note that "entities" don't really exist - they are merely a key with
which a component is associated.